refactor: Replace NameFactory class with helper functions
- Replace NameFactory class with individual helper functions
- Add missing helper functions: buildContextName, buildProviderName, buildServiceName
- Update context-file.ts to use helper functions instead of class instance
- Rename name-factory.ts to name-helpers.ts to better reflect its purpose
- Remove CLAUDE.md from version control (added to .gitignore)

This completes the migration to v0.2.0 patterns using helper functions
instead of classes for better modularity and tree-shaking.
